Conflicts:

	versaplexd/t/schemamatic.t.cs
	versaplexd/vxdbschema.cs

+++ versaplexd/t/vxsanity.t.cs
@@ -125,7 +125,13 @@ class VxSanityTests : VersaplexTester
 	    WVPASSEQ(schema.Length, vxcolinfo.Length);
 
             try {
+	    bool ismono = Wv.wv.IsMono();
 	    for (int colNum = 0; colNum < schema.Length; colNum++) {
+		// Need these for an inconsistency in MS .NET and Mono
+		// conversion of the SQL 'bigint', 'bool', and 'binary' table
+		// types.
+		bool bypass_data_checks = false;
+		bool special_string_size_check = false;
 		var c = schema[colNum];
 
 		WVPASSEQ(c.name, vxcolinfo[colNum].ColumnName);
@@ -133,6 +139,7 @@ class VxSanityTests : VersaplexTester
                 // switch...
                 Type type = c.type;
                 string vxtype = vxcolinfo[colNum].ColumnType.ToLowerInvariant();
+		
                 if (type == typeof(Int64)) {
                     WVPASSEQ(vxtype, "int64");
                 } else if (type == typeof(Int32)) {
@@ -150,21 +157,60 @@ class VxSanityTests : VersaplexTester
                 } else if (type == typeof(Byte[])) {
                     WVPASSEQ(vxtype, "binary");
                 } else if (type == typeof(string)) {
+		    //If we are running Mono for these tests, set this flag
+		    //see below where it's checked...
+		    special_string_size_check = ismono;
                     WVPASSEQ(vxtype, "string");
                 } else if (type == typeof(DateTime)) {
                     WVPASSEQ(vxtype, "datetime");
                 } else if (type == typeof(Decimal)) {
-                    WVPASSEQ(vxtype, "decimal");
+		    // Ugh, OK... so... Mono ends up telling us that a column
+		    // of type 'bigint' is System.Decimal; MS .NET says it's
+		    // a System.Int64.  As a result, this check fails if we
+		    // run the unit testing suite from Linux against a
+		    // versaplexd running on Windows and MS .NET.  Luckily, we
+		    // know how to write hacky code.
+		    if (ismono && colType == "bigint")
+		    {
+			bypass_data_checks = true;
+			WVPASS(vxtype == "decimal" || vxtype == "int64");
+		    }
+		    else
+			WVPASSEQ(vxtype, "decimal");
                 } else {
                     bool return_column_type_is_known = false;
                     WVASSERT(return_column_type_is_known);
                 }
 
-                WVPASSEQ(c.size, vxcolinfo[colNum].Size);
-                // These next two may have problems with mono vs microsoft
-                // differences
-                WVPASSEQ(c.precision, vxcolinfo[colNum].Precision);
-                WVPASSEQ(c.scale, vxcolinfo[colNum].Scale);
+		if (bypass_data_checks)
+		    continue;
+		
+		// This is necessary because Mono's SQL querying functions
+		// return c.size == #characters in a string, whereas MS's
+		// .NET SQL functions return the # of bytes in the string.
+		// String is UTF16 + null character, so multiply the Mono
+		// value by 2 and add 1 to see if it matches the M$ value.
+		// (Of course, if running Versaplex under Mono and the testing
+		// from Mono, c.size == vxcolinfo[colNum].Size)
+		if (special_string_size_check)
+		    WVPASS(c.size == vxcolinfo[colNum].Size ||
+			    c.size * 2 + 1 == vxcolinfo[colNum].Size)
+		else
+		    WVPASSEQ(c.size, vxcolinfo[colNum].Size);
+		
+		// In Mono's SQL querying functions, every data type gets a
+		// 'precision' and 'scale' value of 0, except for "decimal".
+		// The M$ version sets different precision values.  As such,
+		// no point with this comparison if we're running Mono; if
+		// the Versaplex server is *NOT*, then this will fail, and if it
+		// is, then it will succeed.  Blegh.
+		if (ismono && c.precision == 0 && c.scale == 0)
+		    continue;
+		// These next two may have problems with mono vs microsoft
+		// differences... oh you bet they do; see above.
+		//
+		WVPASSEQ(c.precision, vxcolinfo[colNum].Precision);
+		WVPASSEQ(c.scale, vxcolinfo[colNum].Scale);
             }
             } finally {
                 try { VxExec("DROP TABLE test1"); } catch {}
